test relocation
