+++ tests/integration/awslambda/test_lambda_integration.py
@@ -5,7 +5,6 @@ import time
 from unittest.mock import patch
 
 import pytest
-import xmltodict
 from botocore.exceptions import ClientError
 
 from localstack import config
@@ -30,7 +29,6 @@ from .functions import lambda_integration
 from .test_lambda import (
     TEST_LAMBDA_LIBS,
     TEST_LAMBDA_NODEJS_APIGW_502,
-    TEST_LAMBDA_NODEJS_APIGW_INTEGRATION,
     TEST_LAMBDA_PYTHON,
     TEST_LAMBDA_PYTHON_ECHO,
     TEST_LAMBDA_PYTHON_UNHANDLED_ERROR,
@@ -457,37 +455,6 @@ class TestLambdaHttpInvocation:
         assert lambda_resource == content["resource"]
         assert lambda_request_context_path == content["requestContext"]["path"]
         assert lambda_request_context_resource_path == content["requestContext"]["resourcePath"]
-
-    def test_response_headers_invocation_with_apigw(self, create_lambda_function, lambda_client):
-        lambda_name = f"test_lambda_{short_uid()}"
-        lambda_resource = "/api/v1/{proxy+}"
-        lambda_path = "/api/v1/hello/world"
-
-        create_lambda_function(
-            func_name=lambda_name,
-            zip_file=testutil.create_zip_file(
-                TEST_LAMBDA_NODEJS_APIGW_INTEGRATION, get_content=True
-            ),
-            runtime=LAMBDA_RUNTIME_NODEJS14X,
-            handler="apigw_integration.handler",
-        )
-
-        lambda_uri = aws_stack.lambda_function_arn(lambda_name)
-        target_uri = f"arn:aws:apigateway:{aws_stack.get_region()}:lambda:path/2015-03-31/functions/{lambda_uri}/invocations"
-        result = testutil.connect_api_gateway_to_http_with_lambda_proxy(
-            "test_gateway",
-            target_uri,
-            path=lambda_resource,
-            stage_name=TEST_STAGE_NAME,
-        )
-        api_id = result["id"]
-        url = path_based_url(api_id=api_id, stage_name=TEST_STAGE_NAME, path=lambda_path)
-        result = safe_requests.get(url)
-
-        assert result.status_code == 300
-        assert result.headers["Content-Type"] == "application/xml"
-        body = xmltodict.parse(result.content)
-        assert body.get("message") == "completed"
 
     def test_malformed_response_apigw_invocation(self, create_lambda_function, lambda_client):
         lambda_name = f"test_lambda_{short_uid()}"
